Pull request fixes
