+++ app/src/main/java/com/android/wardrobeManager/ui/images/DesignFilterManager.java
@@ -1,16 +1,9 @@
 package com.android.wardrobeManager.ui.images;
 
-import android.content.Context;
-import android.content.res.Resources;
 import android.graphics.Bitmap;
-import android.graphics.Canvas;
-import android.graphics.Paint;
 
-import com.android.wardrobeManager.R;
-import com.android.wardrobeManager.WardrobeManager;
 import com.android.wardrobeManager.ui.images.filters.Default;
-import com.android.wardrobeManager.ui.images.filters.DiagonalDownRight;
-import com.android.wardrobeManager.ui.images.filters.DiagonalUpRight;
+import com.android.wardrobeManager.ui.images.filters.DiagonalStripes;
 import com.android.wardrobeManager.ui.images.filters.HorizontalGradient;
 import com.android.wardrobeManager.ui.images.filters.HorizontalStripes;
 import com.android.wardrobeManager.ui.images.filters.PolkaDots;
@@ -38,8 +31,8 @@ public class DesignFilterManager {
 
     public DesignFilterManager() {
         filters = new HashMap<>();
-        filters.put("diagonal_down_right", new DiagonalDownRight());
-        filters.put("diagonal_up_right", new DiagonalUpRight());
+        filters.put("diagonal_down_right", new DiagonalStripes(30, 45));
+        filters.put("diagonal_up_right", new DiagonalStripes(30, -45));
         filters.put("vertical_gradient", new VerticalGradient());
         filters.put("horizontal_gradient", new HorizontalGradient());
         filters.put("polka_dots", new PolkaDots());
